Stats are now correctly updating between double evictions
